# backend/src/docs/swagger.yaml
basePath: /
definitions:
  request.ReqSignIn:
    properties:
      account_id:
        type: string
      password:
        type: string
    required:
    - account_id
    - password
    type: object
  request.ReqSignUp:
    properties:
      account_id:
        type: string
      auth_code:
        type: string
      nickname:
        type: string
      password:
        type: string
    required:
    - account_id
    - auth_code
    - nickname
    - password
    type: object
  response.ResAuth:
    properties:
      access_token:
        type: string
      access_token_expired_at:
        type: integer
      account_id:
        type: string
      nickname:
        type: string
      refresh_token:
        type: string
      refresh_token_expired_at:
        type: integer
      user_id:
        type: integer
    type: object
  response.ResMemo:
    properties:
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_pinned:
        type: boolean
      latitude:
        type: number
      location_name:
        type: string
      longitude:
        type: number
      rating:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  response.ResMemoList:
    properties:
      memos:
        items:
          $ref: '#/definitions/response.ResMemo'
        type: array
      total:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Daily Memo API Server
  title: Play Daily API
  version: "1.0"
paths:
  /v0.1/auth/signin:
    post:
      consumes:
      - application/json
      description: 로그인 api
      parameters:
      - description: 로그인 요청 데이터
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ReqSignIn'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResAuth'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 로그인 api
      tags:
      - auth
  /v0.1/auth/signup:
    post:
      consumes:
      - application/json
      description: 회원가입 api
      parameters:
      - description: 회원가입 요청 데이터
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.ReqSignUp'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResAuth'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 회원가입 api
      tags:
      - auth
  /v0.1/memo:
    get:
      description: 사용자의 모든 메모를 조회합니다
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResMemoList'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 메모 목록 조회 API
      tags:
      - memo
    post:
      consumes:
      - multipart/form-data
      description: 새로운 메모를 생성합니다 (이미지 파일 포함 가능)
      parameters:
      - description: 메모 제목
        in: formData
        name: title
        required: true
        type: string
      - description: 메모 내용
        in: formData
        name: content
        type: string
      - description: 이미지 파일
        in: formData
        name: image
        type: file
      - description: 평점 (0-5)
        in: formData
        name: rating
        type: integer
      - description: 고정 여부
        in: formData
        name: is_pinned
        type: boolean
      - description: 위도
        in: formData
        name: latitude
        type: number
      - description: 경도
        in: formData
        name: longitude
        type: number
      - description: 장소명
        in: formData
        name: location_name
        type: string
      - description: 카테고리
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.ResMemo'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 메모 생성 API
      tags:
      - memo
  /v0.1/memo/{id}:
    delete:
      description: 메모를 삭제합니다
      parameters:
      - description: 메모 ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 메모 삭제 API
      tags:
      - memo
    get:
      description: 특정 메모를 조회합니다
      parameters:
      - description: 메모 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResMemo'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 메모 조회 API
      tags:
      - memo
    put:
      consumes:
      - multipart/form-data
      description: 메모를 수정합니다 (이미지 파일 포함 가능)
      parameters:
      - description: 메모 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 메모 제목
        in: formData
        name: title
        type: string
      - description: 메모 내용
        in: formData
        name: content
        type: string
      - description: 이미지 파일
        in: formData
        name: image
        type: file
      - description: 평점 (0-5)
        in: formData
        name: rating
        type: integer
      - description: 고정 여부
        in: formData
        name: is_pinned
        type: boolean
      - description: 위도
        in: formData
        name: latitude
        type: number
      - description: 경도
        in: formData
        name: longitude
        type: number
      - description: 장소명
        in: formData
        name: location_name
        type: string
      - description: 카테고리
        in: formData
        name: category
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.ResMemo'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: 메모 수정 API
      tags:
      - memo
swagger: "2.0"